compat: add a bunch of compat code for KDE
Create shadow objects that expose the old 0.8 API just for KDE for the
mean time, while they clean up their network abstraction to work with
NM 0.8 and NM 0.9.  Stuff that talks to the old API should need
minimal changes, basically just changing the interface name for a
few objects, as well as a new dbus object path for the main manager
object.