compat: fix up device hierarchy and properties
Due to dbus-glib limitations, we can't have more than one interface
on the same GObject, so we have to mirror the actual class structure
of NM itself, with a base Device class that implements the .Device interface
and subclasses that implement the .Wired, .Wireless, etc interfaces.

This is also due to the fact that since we're using separate object
paths for the compat objects, instead of having the actual
NMDevice subclasses implement the compat interfaces directly, which
was an approach not taken as the compat code would be much more
intermingled.

We can also simplify property lookups by using g_object_get_property()
instead of explicitly listing out each property, and just override the
ones we need to fix up.