compat: fix up active connection properties change signals
We can't define the same signal name in a superclass and a subclass due
to the way dbus-glib works, since doing so causes glib warnings that
the subclass has defined the same signal as a superclass.  That's fine,
because the properties changed signal, if defined on a subclass,
will work just fine for superclass properties too.

So add a new NMCompatActRequest subclass that will handle the props
changed signal for NMActRequest, like the NMCompatVpnConnection class
does for the VPN stuff, and reorganize things to make it all work.