compat: fix up compat VPN connection specific object property (rh #699786)
The specific-object property for compat VPN connection objects was showing
the active connection path of the actual base ActiveConnection, so we need
to fix that up.

Additionally, the Manager might be given a compat active connection path
as the specific object when activating a VPN connection, so we need to
look up the real NMActRequest that compat path refers to.  That should fix
VPN connection activation problems in KDE.