agents: ensure agent registrations come from a local session
This was already ensured by the at_console section of the dbus policy,
but since we'd like to remove the at_console stuff, we need to handle
local/remote decisions in the code instead of the dbus config file
now.

Agents should only come from local sessions because (at least right
now) it doesn't make a lot of sense for a remote/SSH/VNC/etc user
to provide or save passwords via an agent, since almost all uses
of this functionality may reconfigure the network the user is
logged in over.