dbus: remove at_console policy section
Remove restrictions on most D-Bus methods for remote sessions. This
turns out to be a NOP for any PolicyKit-protected methods because
PK automatically denies non-local users unless the <allow_any>
directive is used, which NM does not use anywhwere.

(If a build disables PolicyKit, then clearly these methods are no
longer protected restricted to local sessions, but the cases where
PK is disabled are likely the same cases where session tracking,
and thus local/remote determination, are also disabled.)

There's no particular reason to deny remote sessions the ability
to inspect network state and read connections visible to that
session.  But those sessions are still denied the ability to
register agents for password handling as there seems to be no
good reason to allow it for now.