core: fix up NMConnectionProvider and NMHostnameProvider interfaces
The code was using "NMConnectionProvider" as the name of both the
interface struct and the instance struct in different places (likewise
with NMHostnameProvider). Fix that by adding "Interface" to the name
of the interface struct, which then also lets us use
G_DEFINE_INTERFACE.

Also, move some gtk-docs from nm-connection-provider.h to
nm-connection-provider.c.

(NMSystemConfigInterface still has the same problem, but is more
awkward to fix since it has "interface" in its name already...)