settings: enforce s_ip4 and s_ip6 for eligible connections
NetworkManager code doesn't need to special-case s_ip4 and s_ip6 any
more and the logic for the defaults is kept in one place (unless
overriden in a specific configuration plugin).

The main motivation for this change was to make nm_connection_compare()
work properly when s_ip4/s_ip6 isn't set by a configuration plugin
without having to special case NULL values.

Note: We're using IPv6 LINK_LOCAL as an equivalent to IPv4 DISABLED. We
should later rename LINK_LOCAL to DISABLED and be done with that.