WIP: dhcp: refactor dhcp manager to support queueing of clients and async killing
Up to now, dhcp clients will be killed synchronously. This is especially
problematic, because in some cases dhclient does not terminate on
SIGTERM, so we block the main loop until timeout.

This change makes it possible to kill the client asynchronously. Now,
when the manager recieves a signal that the client was removed/timed
out, it will no longer remove the client immediatly from its list.

Instead it will keep it using a weak ref. Only when the clients
references all get freed, the manager will remove it from its list.

At the same time, when starting a new client, the manager looks if there
is already a client for the same iface/ipv6 running. If so, it will not
immediatly start the new client. Instead it will queue it.

Killing the client now asynchronously can be done by taking an
additional ref of the client instance, send the remove signal and
wait for the process to terminate asynchronously. When the process
finally terminates, free the ref which in turn will notify the
dhcp-manager that the client is now really disposed. In that case
it might start any queued client.

Signed-off-by: Thomas Haller <thaller@redhat.com>