ifcfg-rh: read and construct iBFT VLAN connections (rh #990480)
dracut/anaconda write out minimal ifcfg files for iBFT connections that
basically just say "use iBFT for the details".  That works OK when the
iBFT interface is a normal ethernet one, but it doesn't work when it's
supposed to be a VLAN because the minimal ifcfg file is not formatted
like a normal ifcfg VLAN file.