core: add NM_UNMANAGED_CHILD and don't assume when set
Instead of ignoring interfaces (like DEVTYPE=wwan) or removing them
when a parent device sets the IP interface (like WWAN PPP or network
interface), continue to expose the child interfaces in the API but
simply unmanage them instead.

Also, if we know an interface is a child interface, don't try to
assume a connection on that interface since the parent is what
controls that child interface's configuration.