tui: rename vpn-helpers and add it to the Makefile
Rename vpn-helpers.c to nm-editor-vpn-helpers.c, update it for various
API changes, and start building it.

To avoid accidentally loading libnm-glib-based VPN plugins (which
would cause an immediate crash), we now look for a different key in
the VPN .name file. (This also means that VPNs will be able to install
both libnm-glib and libnm editor plugins.)