connectivity: explicitly set the default interval for connectivity checking
This matches what is documented in the man page; 300 seconds as the default
value for the connectivity checking interval if it's not set in the config
file or passed as a command-line parameter.

This is already fixed on master by 652853e0d0749a12db07e34f1a3cd901020a76f8.

https://bugzilla.gnome.org/show_bug.cgi?id=723350

[thaller@redhat.com: I modified the original patch]