secrets: call CancelGetSecrets after a timeout in secret request
When a call to GetSecrets fails due to timeout, we have to notify the
agent that we're not interested in the secrets anymore, so that any
pending authentication dialog can show the proper message to user.
This was already the normal behavior before commit 92dda6472cb8
("secret-agent: rework handling of asynchronous request and
cancelling").

The existing code in nm-agent-manager.c which supposedly does this is
never reached and the secret agent seems a better place to implement
the cancellation after a timeout.

Fixes: 92dda6472cb881fc9965a014f16ffabb370b1e55