Fixes to update PacRunner at the right moment.
src/: Fixes in core to update PacRunner at the right place and moment.
When a device goes up PacRunner is configured with the Device IPxConfigs
and Proxy Config. When it goes down the same configuration is removed
from PacRunner.

ifcfg-rh: Fixed to read and write proxy settings to the ifcfg network
scripts.