device: ensure we don't have pending updates before checking for IPv6LL addresses
When checking if a device has already a IPv6LL address, be sure we don't
have related pending events to be processed from platform: if so,
process them immediately. Otherwise we may end skipping IPv6 address
generation because we have not yet processed a "removed address" event:
this will result with a device without any IPv6LL address.

https://bugzilla.redhat.com/show_bug.cgi?id=1452046