libnm: validate NMTCAction.kind and make it a static string
tc_commit() cannot just do
  tfilter->action.kind = nm_tc_action_get_kind (action);
because (previously), the lifetime of nm_tc_action_get_kind()
is not guaranteed to exceed the lifetime of the tfilter
instance. Or at least, it is not obvious, that it does.

tfilter->action.kind can only be set to a static string,
because NMPlatformTfilter has not implemented a way to
manage the lifetime of the strings. For that, it would need
to implement cmd_obj_copy().

But platform anyway needs to craft a netlink message based
on the action kind. The kind is not an opaque string. Its
value must be understood at least by platform to create
the action. Hence, also libnm-core can just as well know the
valid kinds and it can reject invalid ones.

Thus, let's not dynamically allocate the string for kind
but make it a static string. This can be copied around without
danger for the lifeimte of the strings.