all: make simple action's "sdata" string
For kernel, the sdata is a NUL terminated string. Very much like
file names or interface names or basically everything, kernel
is agnostic to any encoding. Of course, that is just lazy and leaves
user space in an impossible situation. User space rightly wants to
treat these fields as strings, show them in a UI, and read them
from command line options (like, the terminal should display
characters according to a certain encoding, not just show �).
But once a text is converted to a bunch of bytes, it only makes
sense if you know the text encoding that is used. But as kernel does
not restrict the encoding in any way (except requiring NUL termination,
and in case of file names special treatment of '.' and '/' -- ASCII
encoded), user space cannot reliably convert the blob back into a string.

Still, overall we don't want to treat such strings as blobs. In most
cases, users are sensible enough to just set them to ASCII or at most
UTF-8 encoding. So, treating them as UTF-8 strings is a good choice.

Change the attribute type of "sdata" from bytestring to string for
that, but make a special exception: this string is not taken literally,
instead it is a backslash encoded UTF-8 string.

That means, in most of the cases, the user can just treat it as
a regular string. Except that he must take special care of backslash.
But it still gives the user the possibility to encode every binary
value in this string using backslash escaping.

The escaped value can be created with nm_utils_str_utf8safe_escape()
or g_strescape() and converted back via nm_utils_str_utf8safe_unescape()
or g_strcompress().

Also, the attributes are not verified. Before accessing them in
tc_commit(), check that the attribute has the expected type.