libnm-core: export a wrapper of json_object_iter_next () function
This is required in order to overcome the clash of the symbol in shared
libraries libjansson and libjson-glib: this happens in gnome-control-center,
which loads the libjson-glib shared library and the libnm one, which has
libjansson as a dependency.
Declare also the macro 'nm_json_object_foreach' and
'nm_json_object_foreach_safe' which are basically a clone of the
'json_object_foreach' and 'json_object_foreach_safe' ones, but drop the
clashing symbol in favor of the new wrapper.