contrib/rpm: don't be too optimistic about enabling fatal warnings
Compilers change and err on the side of caution. That is perfectly well
for those of us who care about programming things defensively, but is
somewhat annoying to those who just wish to build the thing.

We do have a separate debug conditional for enabling extra checks, use
it instead of the test one.