cli: extend printing to support different styles of headers
For
  $ nmcli -f ALL device show wlan0
we print
  AP[1].SSID:                             home

Note how the section header (AP[1]) is not only a static header,
instead it also contains an index (because we might have multiple
APs).

This output format was not supported by nmc_print() before.

Extend nmc_print() and the meta data, to handle displaying such
property types and groups.

There is no change in behavior, because no property is yet marked
as with_indexed_header.