cli: handle field selector prefix in nm_meta_selection_create_parse_list()
The same fields, can be printed with or without a particular prefix,
for example:

  $ nmcli -f AP.SSID device show wlan0

vs.

  $ nmcli -f SSID device wifi

So, there are places, where we must artificially prepend a prefix to the
field selector string.

Move the code to do that to nm_meta_selection_create_parse_list().

A question that arises is, if the "fields_str" is intended to be used
without prefix, why do we pass a nested fields array, like

   NMC_META_GENERIC_GROUP ("GENERAL", metagen_con_active_general, N_("GROUP"))

The answer to that is, that we need the nesting (and the hierarchical
information) during nmc_print(), to properly show the fields.
However, for selecting, fields_str should be specified without
the outer selector.