cli: extend processing of fields selection for common/all types
It's complicated. The function that parses the fields list and selects
the relevant bits out of a list of NMMetaAbstractInfo entries,
gets extended to natively understand the "all" and "common" fields.

Up to now, "all" and "common" was handled by preprocessing the fields
string and replace "all" and "common" with appropriate list of fields
(in case of "all", %NULL was used, which mean to select all fields).

However, that does not work with nested fields, and if we want to nest
fields deeper than one level. If you have a "common" selector, we must
select common fields recursively. Same with "all". Hence, these special
fields must be handled by the parsing code itself.

Also, previously it was not possible to select -f "SOME_GROUP.COMMON" or
-f "SOME_GROUP.ALL". This is still not supported, however it becomes
feasable.