libnm-core: sink GVariants from nm_utils_parse_variant_attributes()
nm_utils_parse_variant_attributes() currently returns a GHashTable
containing floating GVariants, which are unreferenced when the hash
table is disposed. This API is inconvenient because:

 - if the caller sinks the variants, the floating reference will be
   converted to a full one, but this will not prevent the GVariant
   from being disposed when the hash table is destroyed.

 - if the caller refs the variants, this will add a new reference in
   addition to the floating one, which is confusing because any
   following g_variant_ref_sink() will probably do the wrong thing.

In other words the problem is that if the hash table owns the variant,
it should have a full reference to it. Fix this and improve the
function documentation.

Note that this changes behavior of a libnm public API. After
discussion we decided this is acceptable because the previous behavior
was a bug, and the documentation didn't specify that the returned
variants were floating. If anybody is using the function and upgrades
to a libnm version with this fix, the full reference instead of the
floating one could cause some leaks if the code did some assumptions
not justified by the function documentation.

https://bugzilla.redhat.com/show_bug.cgi?id=1594887