policy: track best active connections rather than best devices
NMManager exports a PrimaryConnection property which gets updated when
the default_ipX_device property of NMPolicy changes.  The primary
connection is set to the activation request currently pending on the
best default device. The default device in NMPolicy is determined by
evaluating all VPNs (first choice) and devices (otherwise) and
selecting the one with the best default route.

Selecting a best device to determine a primary connection is ambiguous
because in presence of VPNs there are multiple active connections (the
parent connection and the VPN) for the same device. We currently set
the parent connection as primary one even if the VPN has the best
default route.

It seems a better choice to track best active connections directly
instead from NMPolicy to avoid one lossy intermediate step when
determining the primary connection.