manager: treat VPN generated connections as managed
We create in-memory connections for VPN virtual devices as tun and we
keep them as external, which means that NM doesn't touch them in any
way. This is a fake promise, as NM indirectly created them through the
VPN activation. Also, this means that NM can't commit new routes to
the device; routes that could be needed if we want to activate another
VPN over the existing one. So, we should fully manage those devices.

Note: there are still some issues to fix, as fully managing the
devices creates duplicate routes at the moment.