fixup! core: introduce NMDeviceWireguard
The private key is not exported on D-Bus (see
introspection/org.freedesktop.NetworkManager.Device.Wireguard.xml).

I think that is right, and it should not be exported.

Hence, this property is unused, because GObject properties under
"src/" are are usually only used to glue D-Bus exported properties
to gdbus (via NMDBusObject).