wifi: no longer special case SSID "<hidden>"
We coerced SSIDs with value "<hidden>" or "<hidden>\0" to
an empty SSID to indicate hidden APIs.

See commits:

 - cd475e6d0cba05e152ce0f25d72448e17407ac30: 2004-12-15

   /* NOTE: some Cisco products actually broadcast "<hidden>" as their ESSID when they
    * are set to not broadcast it, rather than just broadcasting a blank ESSID.
    */

 - ff6dbffeb1cc1bd23cf5595fdc0c98567ff22360: 2005-01-07

   /* ipw2x00 drivers fill in an essid of "<hidden>" if they think the access point
    * is hiding its MAC address.  Sigh.

 - bbbd00d4c5af7a6342df8be75cc4ca177feabecf: 2005-02-27

   /* Stupid ipw drivers use <hidden> */

 - 56621e5f66e1e6285563ff0a29af485f15ccc978: 2006-11-26

   /* Stupid ieee80211 layer uses <hidden> */

I think this is wrong, and I doubt it is still necessary. First of all, hidden
APs are hardly relevant anymore. Second, it is unclear, whether this is a driver
problem or the AP.

For NetworkManager, the SSID is an opaque blob. It has no business in interpreting
certain SSIDs to mean something else. If this is a driver/supplicant issue, the lower
layers should be fixed. If this is an AP issue, we should evaluate whether such APs
are still in use. Also, if the AP is hidden, that does not mean that it requires the
AP should announce a particular SSID "<hidden>".

This code is likely of no relevance anymore. Drop it.

If at all, such workarounds belong to the lower layers, either
supplicant or the driver.