libnm: use nm_utils_fd_get_contents() for nm_vpn_service_plugin_read_vpn_details()
There is an important difference: if an IO error happens, the previous
implementation would scrape whatever it could read so far. That differs
from nm_utils_fd_get_contents() which would fail altogether. Whether
that is desirable or not, is a question.

If that is a problem, we should instead add a flag to
nm_utils_fd_get_contents() for returning the data that was
read so far.

Other than that, there are two upsides:

- nm_utils_fd_get_contents() and NMSecretPtr ensure that there
  are no secrets left in temporary memory.
- nm_utils_fd_get_contents() is our best-practice implementation
  of how to read from a file descriptor. We should not reimplement
  this functionality, because there are many subtleties that we
  solved already at one place.