shared: add NMStrBuf util
Our own implementation of a string buffer like GString.

The most important difference is, that it can optionally
nm_explicit_bzero() the buffers while resizing. If we want
to avoid leaking secrets while constructing a string,
we cannot use GString for that.

Also, GString is heap allocated, for little reason.

It's surprisingly simple to provide such an implementation.
Do it all header-only, so that it can be possibly inlined.