cli/devices: sort hotspots above in nmcli output
In general we want to keep the connections that the user is most likely
to want to see topost. Default connections should be close to the top,
but the connections that are likely to have been brought up manually
shall be above them. It applies to VPN connections and should apply to
Hotspots too.