libnm/team: update JSON config when using accessor functions
When one of the libnm accessors (like nm_setting_team_add_runner_tx_hash())
is called, then we *MUST* update the JSON as well.

Well, technically there are two options:

(1) setting one of the artifical properties does not modify the JSON at
all. On D-Bus, this would send the properties like runner-rx-hash, and
a new JSON would be constructed from that.

(2) when one of the artificial properties is modified, then the JSON
must be updated.

Note that set_property() already does (2) for all properties. The direct
libnm accessors must do that as well.

This is a serious problem, and only mitigated because nmcli works around
that [1] by first removing all entries, then adding them. First of all,
set_fcn() is not supposed to replace all values, it's supposed to add
them.
Anyway, this this only appeared to work, because nmcli would only modify
priv->runner_tx_hash, which was sent via D-Bus. But on D-Bus "config" would
still be invalid. The server then would re-assemble the setting, first
parsing "config" and the overwriting it with the received
"runner-tx-hash" setting. But this is not right and only works in this
special case.

In general, setting a property via libnm API must update the JSON right
away, like done for all GObject setters.

[1] https://cgit.freedesktop.org/NetworkManager/NetworkManager/tree/clients/common/nm-meta-setting-desc.c?id=cfcd746260b0dc1af80dd769d67dfa2649569b36#n4082