libnm/team: don't set properties explicitly in set_property()
As we are anyway going to call _align_team_properties(), there is
no point in setting the value first. At best, it's duplicate
and unnecessary work. At worst, the implementations don't actually
do the same.

Note that _align_team_properties() eventually calls
_nm_utils_team_config_get(). Thus, if libnm is compiled without
libjansson, then this will always yield the default value
and setting the property shall have no effect. That is desired,
since, JSON support is not available.
Note that this is no change to before. Also before we would always
call _align_team_properties() and set the default. We just skip
setting the value first to something else.

Also skip the check whether we are setting the same value.
It requires us to be sure that the GValue contains the same
after we would do the JSON round trip. That is non-obvious.
If you are concerned about the additional overhead, there are other
places to optimize.