vpn: set STOPPED state when service disappears
When the VPN service daemon fails it sends a state-change signal to NM
and quits; however it is difficult to guarantee that the signal is
delivered to NM before the process termination and so in some cases
the signal is lost. Work around that by manually setting the STOPPED
state when the service disappears from the bus.